+++ compressors/baseline_compressors.py
@@ -5,7 +5,7 @@ Contains some elementary baseline compressors
 
 from core.data_block import DataBlock
 from core.encoded_stream import EncodedBlockReader, EncodedBlockWriter
-from core.framework import DataEncoder, DataDecoder
+from core.data_encoder_decoder import DataEncoder, DataDecoder
 from utils.bitarray_utils import BitArray, bitarray_to_uint, uint_to_bitarray, get_bit_width
 from utils.test_utils import try_lossless_compression
 import tempfile
