+++ compressors/integer_coders.py
@@ -1,5 +1,5 @@
 from core.data_block import DataBlock
-from core.framework import DataEncoder, DataDecoder
+from core.data_encoder_decoder import DataEncoder, DataDecoder
 from utils.bitarray_utils import uint_to_bitarray, bitarray_to_uint, BitArray
 from utils.test_utils import try_lossless_compression
