+++ compressors/prefix_free_compressors.py
@@ -3,7 +3,7 @@ from typing import Mapping
 from utils.bitarray_utils import BitArray
 from utils.tree_utils import BinaryNode
 from core.prob_dist import ProbabilityDist
-from core.framework import DataEncoder, DataDecoder
+from core.data_encoder_decoder import DataEncoder, DataDecoder
 from core.data_block import DataBlock
