+++ core/data_block.py
@@ -1,5 +1,4 @@
 from typing import List, Set
-from utils.bitarray_utils import compute_alphabet, compute_counts_dict
 from core.prob_dist import ProbabilityDist
 import unittest
